File restructure & Grid tile
Added an example card resource (wywy swordsman).

Restructured which scene will appear as the level. I have decided that there will only be one level scene for now, and all the variations will in the parameters (e.g. whether or not opponent is AI, background image, grid size, etc.)

Implemented a basic grid tile where I can place a card. The animation is not correct and it doesn't do anything important, but at least it's a start! Also, the nodes containing all the cards are not correct. This kinda sucks, but it's OK.