Level UI & Grid Resize
I made the important decision today to reduce the grid size to a width of 5 and a height of 4. This is because the UI height is too small, and I want to keep the game easy on the eyes.

The control node migration is hitting really hard---almost everything is a control node now!