Basic Attack Framework
The code relating to this commit is bugged, has not been tested, and is incomplete. I'm pushing this because this is the end of my one-day CS challenge where I attempted to code card attack logic in a single day. As you can see here, I failed, but not by a super long shot.