Add New Cards, Loss Condition, Fix Bugs
Fixed a bug where cards would target themselves. Cards now should target the correct enemy (lol flipping coords when it's the server is a little wild, we should do that on the non-server clients :P)

Added two new cards, the Nebula Trooper for the wywys, and the Giant Stickman for the Stickmen.

Removed the export attribute from the Card class. It should've never been there because images are rendered when the cardID is being updated.

Added the ability for players to lose health from direct attacks.