Add Specials
Made the first player always the Wywy and the second player always the Stickman.

Added MoveLane.

Added Specials.

Added Whey-Gas.

Added docstrings to Card and Move.

Added a findTarget function in the Lobby autoload. Hopefully this will make it easier to find a valid target.

Moved attack bonuses members from Entity to Card.

Fixed a typo when deserializing the defense multiplier in Entity.